requirements.txt
django~=4.1
django-crispy-forms~=1.14.0
pymysql~=1.0.2